+++ src/lib/plugins/value/components/ReadonlyValue.svelte
@@ -43,6 +43,7 @@
   }
 </script>
 
+<!-- svelte-ignore a11y-click-events-have-key-events -->
 <div
   data-type="selectable-value"
   class={getValueClass(value, parser)}