+++ src/lib/types.ts
@@ -11,6 +11,11 @@ export type Content = JSONContent | TextContent
 
 export type JSONParser = JSON
 
+export interface JSONPathParser {
+  parse: (pathStr) => JSONPath
+  stringify: (path: JSONPath) => string
+}
+
 export interface VisibleSection {
   start: number
   end: number
@@ -113,7 +118,7 @@ export type ClipboardValues = Array<{ key: string; value: JSONValue }>
 export interface FontAwesomeIcon {
   prefix: string
   iconName: string
-  icon: [number, number, Array<number | string>, string, string]
+  icon: [number, number, Array<number | string>, string | null, string]
 }
 
 export interface DropdownButtonItem {