+++ src/routes/examples/custom_menu_buttons/+page.svelte
@@ -4,7 +4,7 @@
   import copyToClipboard from '$lib/utils/copyToClipboard'
 
   let content = {
-    text: undefined, // used when in text mode
+    text: undefined, // can be used to pass a stringified JSON document instead
     json: {
       array: [1, 2, 3],
       boolean: true,