+++ src/routes/examples/use_methods/+page.svelte
@@ -4,7 +4,7 @@
   let refJsonEditor
 
   let content = {
-    text: undefined, // used when in text mode
+    text: undefined, // can be used to pass a stringified JSON document instead
     json: {
       array: [1, 2, 3],
       boolean: true,